agents/requirements.txt
# Core Web Framework
fastapi==0.118.0
uvicorn==0.37.0
python-multipart==0.0.20
jinja2==3.1.6

# AI and Machine Learning
crewai==0.201.1
crewai-tools==0.75.0
cerebras_cloud_sdk==1.50.1
google-generativeai==0.8.5

# Data Processing
pandas==2.3.3
PyPDF2==3.0.1
pydantic==2.11.9

# AWS and Cloud Storage
boto3==1.40.45
botocore==1.40.45

# Security and Encryption
cryptography==46.0.1

# Environment and Configuration
python-dotenv==1.1.1
PyYAML==6.0.3

# Text Processing
markdown==3.9

# HTTP Requests
requests==2.32.5

# Utility Libraries
pathlib==1.0.1
typing_extensions==4.15.0

# Dependencies (automatically installed but listed for clarity)
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
idna==3.10
MarkupSafe==3.0.3
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
urllib3==2.3.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
